Consolidate the request-field validation helpers into one header
The presence/validation helpers had accreted across two changes into
require_scalar.h (which despite its name also held enum/int/bool guards) and a
separate require_period.h. They are now one require_field.h with a single
consistent doc-comment covering the scalar, enum, int, bool and Period guards.
Function names and behavior are unchanged, so the 31 call sites only swap their
include. Every consumer is already a FlatBuffers/QuantLib translation unit, so
folding the Period header's includes in costs nothing.

Full suite green (682 cases, unchanged).
