+++ src/mappers/basis_swap_mapper.cpp
@@ -6,7 +6,7 @@
 #include "enum_convert.h"
 #include "error.h"
 #include "leg_notionals.h"
-#include "require_scalar.h"
+#include "require_field.h"
 
 namespace quantra {
