+++ src/mappers/bootstrap_inflation_curves_mapper.cpp
@@ -6,7 +6,7 @@
 
 #include "date_convert.h"
 #include "enum_convert.h"
-#include "require_scalar.h"
+#include "require_field.h"
 #include "error.h"
 #include "grid_utils.h"
