+++ src/mappers/calibrate_swaption_model_mapper.cpp
@@ -3,7 +3,7 @@
 #include "enum_convert.h"
 #include "error.h"
 #include "model_generated.h"
-#include "require_period.h"
+#include "require_field.h"
 
 namespace quantra {
