+++ src/mappers/cap_floor_mapper.cpp
@@ -4,7 +4,7 @@
 #include "schedule_parser.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_scalar.h"
+#include "require_field.h"
 
 namespace quantra {
