+++ src/mappers/cds_mapper.cpp
@@ -6,7 +6,7 @@
 #include "schedule_parser.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_scalar.h"
+#include "require_field.h"
 
 namespace quantra {
