+++ src/mappers/fra_mapper.cpp
@@ -3,7 +3,7 @@
 #include "date_convert.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_scalar.h"
+#include "require_field.h"
 
 namespace quantra {
