+++ src/mappers/sample_vol_surfaces_mapper.cpp
@@ -4,7 +4,7 @@
 
 #include "date_convert.h"
 #include "enum_convert.h"
-#include "require_scalar.h"
+#include "require_field.h"
 #include "error.h"
 #include "swaption_vol_diagnostics.h"
