+++ src/mappers/year_on_year_inflation_cap_floor_mapper.cpp
@@ -4,8 +4,7 @@
 #include "enum_convert.h"
 #include "schedule_parser.h"
 #include "error.h"
-#include "require_scalar.h"
-#include "require_period.h"
+#include "require_field.h"
 
 namespace quantra {
