+++ src/mappers/year_on_year_inflation_swap_mapper.cpp
@@ -6,8 +6,7 @@
 #include "schedule_parser.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_scalar.h"
-#include "require_period.h"
+#include "require_field.h"
 
 namespace quantra {
