+++ src/mappers/zero_coupon_inflation_swap_mapper.cpp
@@ -5,8 +5,7 @@
 #include "date_convert.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_scalar.h"
-#include "require_period.h"
+#include "require_field.h"
 
 namespace quantra {
