+++ src/parsers/equity_option_parser.cpp
@@ -6,7 +6,7 @@
 #include "date_convert.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_scalar.h"
+#include "require_field.h"
 
 namespace quantra {
