+++ src/parsers/inflation_curve_parsers.cpp
@@ -22,8 +22,7 @@
 #include "date_convert.h"
 #include "enum_convert.h"
 #include "error.h"
-#include "require_scalar.h"
-#include "require_period.h"
+#include "require_field.h"
 #include "index_registry_builder.h" // CurrencyFromString
 
 namespace quantra {
