+++ src/parsers/term_structure_parser.cpp
@@ -9,8 +9,7 @@
 
 #include <cmath>
 
-#include "require_scalar.h"
-#include "require_period.h"
+#include "require_field.h"
 
 using namespace QuantLib;
