+++ src/parsers/zero_coupon_bond_parser.cpp
@@ -1,7 +1,7 @@
 #include "zero_coupon_bond_parser.h"
 
 #include "error.h"
-#include "require_scalar.h"
+#include "require_field.h"
 
 std::shared_ptr<QuantLib::Bond> ZeroCouponBondParser::parse(const quantra::ZeroCouponBond *bond)
 {
